% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage information from an insurance company}
\format{
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{distance from fire station in miles}
  \item{DAMAGE}{x 1000 dollars in damage}
  ...
}
}
\source{
Data from \url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
Fire Damage information from an insurance company
}
\keyword{datasets}
