% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{a binomial distribution where user gets to select the number of iterations ran}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{number of times the function will run}

\item{n}{the number of in a trial}

\item{p}{the probablity of getting a success}
}
\value{
a boxplot of the data distribution as well as a table of probablities from the run
}
\description{
a binomial distribution where user gets to select the number of iterations ran
}
\examples{
mybin(iter=100, n = 10, p= 0.7)
}
