% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{A bootstraping function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{values used for bootstrapping}

\item{fun}{function used for bootstrapping}

\item{alpha}{alpha value}

\item{cx}{location of text}

\item{...}{addition input for para function to make the histogram}
}
\value{
a histogram showing the function's values after bootstrapping
}
\description{
A bootstraping function
}
\examples{
myboot2(x=c(1,1,1,2,2,2,2,3,3,3,4,4), fun ="median")
}
