% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{A function that will return a 95\% mean confidence interval of one data set}
\usage{
myci(x)
}
\arguments{
\item{x}{vector used}
}
\value{
the 95\% confidence interval for the mean
}
\description{
A function that will return a 95\% mean confidence interval of one data set
}
\examples{
myci(c(1,4,5,2,4))
}
