% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Estimating the CLT using a normal distribution}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{number of values in each iteration}

\item{iter}{the number of times the distribution will be created}

\item{a}{the lower number for the uniform distribution}

\item{b}{the upper number for the uniform distribution}
}
\value{
a plot showing the sample mean vs the relative density of each mean
}
\description{
Estimating the CLT using a normal distribution
}
\examples{
mycltu(20,10,4,8)
}
