% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Finds the maximum likelihood estimate of data provided using the function specified}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{the log function that will be used in calculation}

\item{x}{the data used in calculation}

\item{param}{the number range the data could fall into}

\item{...}{additional options to be placed into the plot command}
}
\value{
a graphical representation of the data as well as a list containing important information
}
\description{
Finds the maximum likelihood estimate of data provided using the function specified
}
\examples{
mymaxlik(x=c(4,6,5),param=seq(0,20,length=1000),lfun=function(x,param) log(dpois(x,lambda=param)))
}
