% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{A normal curve where P(x<=a)}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{Probablitiy that x = a, where P(X<=a)}

\item{mu}{mean of normal distribution}

\item{sigma}{standard deviation of normal distribution}
}
\value{
a plot showing the normal curve, with the probability of x <= a in red
}
\description{
A normal curve where P(x<=a)
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
