% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myquadplot.R
\name{myquadplot}
\alias{myquadplot}
\title{The quadratic line of best fit of Breast Height Diameters to tree heights}
\usage{
myquadplot(x)
}
\arguments{
\item{x}{the variable of Breast Height Diameters of Spruce Trees}
}
\value{
the estimated height of the Spruce Tree
}
\description{
The quadratic line of best fit of Breast Height Diameters to tree heights
}
\examples{
myquadplot(19)
}
