% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Number of tickets to solve an airline overbooking problem}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.94)
}
\arguments{
\item{N}{Number of seats on the plane.}

\item{gamma}{Pain threshold deemed suitable by the airline}

\item{p}{probablility a person shows}
}
\value{
two plots, one showing the discrete number of seats sold and one showing the continuous number of seats sold, as well as list containing useful data from the function
}
\description{
Number of tickets to solve an airline overbooking problem
}
\examples{
ntickets(N=400,gamma = 0.02, p = 0.95)

}
