% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{The mean and standard deviation of a normal distribution based on a normal distribution}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{the sample vector}

\item{mu}{the range the mean of the sample could fall in}

\item{sig}{the range the standard deviation of the sample could fall in}

\item{...}{additional information you can add to the graph}
}
\value{
a list of important variables from this function as well as a plot of found values of mu and sigma
}
\description{
The mean and standard deviation of a normal distribution based on a normal distribution
}
