% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{piecewise}
\alias{piecewise}
\title{Piecewise function at x = 18}
\usage{
piecewise(x, coef)
}
\arguments{
\item{x}{the variable in the equation}

\item{coef}{a coefficent from the data summary}
}
\value{
a one line piecewise function where the break is at x = 18
}
\description{
Piecewise function at x = 18
}
\examples{
\dontrun{myf(x,coef=tmp$coefficients[,"Estimate"])}
}
